Cleanup

+++ services/datamanager/sync.go
@@ -84,11 +84,16 @@ func (s *syncer) upload(ctx context.Context, path string) {
 	s.backgroundWorkers.Add(1)
 	goutils.PanicCapturingGo(func() {
 		defer s.backgroundWorkers.Done()
-		exponentialRetry(
+		uploadErr := exponentialRetry(
 			ctx,
 			func(ctx context.Context) error { return s.uploadFn(ctx, nil, path) },
 			s.logger,
 		)
+		if uploadErr != nil {
+			return
+		}
+
+		// Delete the file and indicate that the upload is done.
 		if err := os.Remove(path); err != nil {
 			s.logger.Errorw("error while deleting file", "error", err)
 		} else {
@@ -130,10 +135,10 @@ func (p *progressTracker) unmark(k string) {
 
 // exponentialRetry calls fn, logs any errors, and retries with exponentially increasing waits from initialWait to a
 // maximum of maxRetryInterval.
-func exponentialRetry(cancelCtx context.Context, fn func(cancelCtx context.Context) error, log golog.Logger) {
+func exponentialRetry(cancelCtx context.Context, fn func(cancelCtx context.Context) error, log golog.Logger) error {
 	// Only create a ticker and enter the retry loop if we actually need to retry.
 	if err := fn(cancelCtx); err == nil {
-		return
+		return nil
 	}
 
 	// First call failed, so begin exponentialRetry with a factor of retryExponentialFactor
@@ -144,13 +149,13 @@ func exponentialRetry(cancelCtx context.Context, fn func(cancelCtx context.Conte
 			if !errors.Is(err, context.Canceled) {
 				log.Errorw("context closed unexpectedly", "error", err)
 			}
-			return
+			return err
 		}
 		select {
 		// If cancelled, return nil.
 		case <-cancelCtx.Done():
 			ticker.Stop()
-			return
+			return cancelCtx.Err()
 		// Otherwise, try again after nextWait.
 		case <-ticker.C:
 			if err := fn(cancelCtx); err != nil {
@@ -163,7 +168,7 @@ func exponentialRetry(cancelCtx context.Context, fn func(cancelCtx context.Conte
 			}
 			// If no error, return.
 			ticker.Stop()
-			return
+			return nil
 		}
 	}
 }
